327. xdm warning fix (Marc La France).  326. Make x11perf dies gracefully when display cannot be opened       (Marc La France).  325. Fix rendition driver compile glitch (Marc La France).  324. #ifdef out glitch in glint driver (Marc La France).  323. Preliminary DGA support in ATI driver (Marc La France).  322. Minor update to DESIGN and Status documents (Marc La France).  321. Fix DGA's reporting of vertical refresh rates to clients       (Marc La France).  320. Make driver-provided Sync(), SetViewport() and GetViewport() DGA       callbacks optional (Marc La France).  319. Fix XKB compile glitch (Marc La France).  318. Make imstt driver compile (Marc La France).

+++ lib/xkbfile/XKBconfig.h
@@ -27,7 +27,7 @@
  THE USE OR PERFORMANCE OF THIS SOFTWARE.
 
  ********************************************************/
-/* $XFree86$ */
+/* $XFree86: xc/lib/xkbfile/XKBconfig.h,v 3.3 2000/06/14 18:20:33 dawes Exp $ */
 
 
 typedef struct _XkbConfigRtrn	*XkbConfigRtrnPtr;
@@ -121,13 +121,14 @@ typedef struct _XkbConfigRtrn {
 	char *			symbols;
 	char *			types;
 	char *			compat;
+
 	Bool			replace_initial_ctrls;
 	unsigned long		initial_ctrls;
 	unsigned long		initial_ctrls_clear;
 
 	Bool			replace_initial_opts;
-	unsigned long		initial_opts;
-	unsigned long		initial_opts_clear;
+	unsigned short		initial_opts;
+	unsigned short		initial_opts_clear;
 
 	XkbConfigModInfoRec	initial_mods;
 	XkbConfigModInfoRec	internal_mods;
